backend/requirements.txt
Flask
flask-cors
paho-mqtt
pymemcache
